refactor(core): drop compatibility re-exports from command_executor
_MAX_RETRY_COUNT, _UNOWNED_WARN_RATE_LIMIT_SECS, and RetryableBatch were
re-exported from execution_pipeline solely so two test files could keep
importing them from the old module path after the pipeline split. Migrate
those tests to import directly from execution_pipeline and drop the shim.
